Fixed the problems
You had separate functions for time and weather and each of them was calling the home function separately. The main change I did was make a new function setUpEverything which did both of the things in one function and called the home page with all the data. I did one more thing in the choose_location, instead of a pushNamed navigation, I did a pushReplacement navigation since the data with the pop was giving problems. If you have any problems, you know my discord